Add cached pinNumber calculation for Pin-Class
Added a cached calculation for the pin represented by an instance of `Pin`. Changing the mask of a `Pin`-instance should not happen that often, so I added the calculation of the number directly after changing the mask.
This is helpful if one wants to use counters, because they need the number of the pin instead of their mask.