Add initial version of timeout-reading
Added an initial version of reading with a timeout functionality. (only one byte for now)
This is done by polling for a start-bit instead of using the hardware-functionality. For a nice and fast way of io-polling with a timeout, I used Phil's approach from here: http://forums.parallax.com/discussion/comment/774604#Comment_774604